del next in Serv, add if check id

+++ controllers/contactsControllers.js
@@ -10,9 +10,13 @@ export const getAllContacts = async (req, res) => {
 };
 
 export const getOneContact = async (req, res) => {
-  try {
-    const { id } = req.params;
+  const { id } = req.params;
 
+  if (!isValidObjectId(id)) {
+    return res.status(404).json({ message: "This identifier is not valid" });
+  }
+
+  try {
     const contact = await contactsService.getContactById(id);
     if (contact) {
       res.status(200).json(contact);
@@ -25,8 +29,13 @@ export const getOneContact = async (req, res) => {
 };
 
 export const deleteContact = async (req, res) => {
+  const { id } = req.params;
+
+  if (!isValidObjectId(id)) {
+    return res.status(404).json({ message: "This identifier is not valid" });
+  }
+
   try {
-    const { id } = req.params;
     const deletedContact = await contactsService.removeContact(id);
     if (deletedContact) {
       res.status(200).json(deletedContact);
@@ -49,8 +58,13 @@ export const createContact = async (req, res) => {
 };
 
 export const updateContact = async (req, res) => {
+  const { id } = req.params;
+
+  if (!isValidObjectId(id)) {
+    return res.status(404).json({ message: "This identifier is not valid" });
+  }
+
   try {
-    const { id } = req.params;
     const contact = await contactsService.getContactById(id);
     if (contact) {
       const { name, email, phone, favorite } = req.body;
@@ -75,6 +89,10 @@ export const updateContact = async (req, res) => {
 export const updateContactFavorite = async (req, res) => {
   const { id } = req.params;
   const { favorite } = req.body;
+
+  if (!isValidObjectId(id)) {
+    return res.status(404).json({ message: "This identifier is not valid" });
+  }
 
   try {
     const updatedContact = await contactsService.updateStatusContact(
